writing cache storing

+++ cachemanager/src/main/java/com/example/alexeyglushkov/cachemanager/DiskCacheProvider.java
@@ -1,18 +1,13 @@
 package com.example.alexeyglushkov.cachemanager;
 
-import com.example.alexeyglushkov.streamlib.InputStreamReader;
-import com.example.alexeyglushkov.streamlib.OutputStreamWriter;
-import com.noveogroup.android.cache.disk.DiskCache;
-import com.noveogroup.android.cache.disk.MetaData;
-import com.noveogroup.android.cache.io.Serializer;
+import com.example.alexeyglushkov.streamlib.Serializer;
 
-import java.io.BufferedOutputStream;
+import java.io.BufferedInputStream;
 import java.io.File;
-import java.io.FileOutputStream;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
-import java.util.Date;
+import java.io.Serializable;
 
 /**
  * Created by alexeyglushkov on 26.09.15.
@@ -20,20 +15,19 @@ import java.util.Date;
 public class DiskCacheProvider implements CacheProvider {
 
     private File directory;
-    private InputStreamReader streamReader;
-    private OutputStreamWriter streamWriter;
+    private Serializer serializer;
+    private Error lastError;
 
-    public DiskCacheProvider(File directory, InputStreamReader streamReader, OutputStreamWriter streamWriter) {
+    public DiskCacheProvider(File directory, Serializer serializer) {
         this.directory = directory;
-        this.streamReader = streamReader;
-        this.streamWriter = streamWriter;
+        this.serializer = serializer;
     }
 
     @Override
-    public Error store(String key, CacheEntry entry, CacheMetadata metadata) {
+    public Error store(String key, Object entry, Serializable metadata) {
         Error error = prepareDirectory();
         if (error == null) {
-            error = write(key, entry, metadata);
+            error = write(key, entry, (DiskCacheMetadata)metadata);
         }
 
         return error;
@@ -51,19 +45,29 @@ public class DiskCacheProvider implements CacheProvider {
         return null;
     }
 
-    private File getKeyDirectory(String key) {
+    private File getKeyFile(String key) {
         String fileName = Integer.toString(key.hashCode());
         File file = new File(directory.getPath() + File.pathSeparator + fileName);
         return file;
     }
 
-    private Error write(String key, CacheEntry entry, CacheMetadata metadata) {
-        Error error = null;
-        File file = getKeyDirectory(key);
+    private File getKeyMetadataFile(String key) {
+        String fileName = Integer.toString(key.hashCode()) + "_metadata";
+        File file = new File(directory.getPath() + File.pathSeparator + fileName);
+        return file;
+    }
 
-        if (file.exists()) {
-            if(!file.delete()) {
-                error = new Error("DiskCacheProvider write delete: can't delete cache file");
+    private Error write(String key, Object object, DiskCacheMetadata metadata) {
+        Error error = null;
+        error = writeMetadata(key, metadata);
+
+        File file = null;
+        if (error == null) {
+            file = getKeyFile(key);
+            if (file.exists()) {
+                if (!file.delete()) {
+                    error = new Error("DiskCacheProvider write delete: can't delete cache file");
+                }
             }
         }
 
@@ -76,44 +80,51 @@ public class DiskCacheProvider implements CacheProvider {
         }
 
         if (error == null) {
-            OutputStream fos = null;
-            try {
-                fos = new BufferedOutputStream(new FileOutputStream(file));
-                error = streamWriter.writeStream(fos, entry);
-            } catch (IOException ex) {
-                error = new Error("DiskCacheProvider write open stream exception:" + ex.getMessage());
-            } finally {
-                if (fos != null) {
-                    try {
-                        fos.close();
-                    } catch (Exception ex) {
-                        error = new Error("DiskCacheProvider write close stream exception:" + ex.getMessage());
-                    }
-                }
-            }
+            DiskCacheEntry entry = new DiskCacheEntry(file, metadata, serializer);
+            error = entry.write();
         }
 
         return error;
     }
 
+    private Error writeMetadata(String key, DiskCacheMetadata metadata) {
+
+    }
+
     @Override
-    public CacheEntry getEntry(String key) {
-        CacheEntry entry = null;
-        File file = getKeyDirectory(key);
+    public Object getValue(String key) {
+        Object entry = null;
+        File file = getKeyFile(key);
 
         if (file.exists()) {
-            InputStream inputStream
-            entry = streamReader.readStream();
+            lastError = new Error("DiskCacheProvider getValue: file doesn't exist");
+        }
+
+        if (lastError == null) {
+            InputStream inputStream = null;
+            try {
+                inputStream = new BufferedInputStream(new FileInputStream(file));
+                entry = (DiskCacheEntry)serializer.read(inputStream);
+            } catch (Exception ex) {
+                lastError = new Error("DiskCacheProvider getValue exception: " + ex.getMessage());
+            } finally {
+
+            }
         }
 
         return entry;
     }
 
-    public CacheMetadata getMetadata(String key) {
+    public Serializable getMetadata(String key) {
         return null;
     }
 
     @Override
     public void remove(String key) {
+    }
+
+    @Override
+    public Error getError() {
+        return lastError;
     }
 }
